+++ ts/src/coder/spl-token/index.ts
@@ -4,6 +4,7 @@ import { SplTokenInstructionCoder } from "./instruction.js";
 import { SplTokenStateCoder } from "./state.js";
 import { SplTokenAccountsCoder } from "./accounts.js";
 import { SplTokenEventsCoder } from "./events.js";
+import { SplTokenTypesCoder } from "./types.js";
 
 /**
  * Coder for the SPL token program.
@@ -13,11 +14,13 @@ export class SplTokenCoder implements Coder {
   readonly accounts: SplTokenAccountsCoder;
   readonly state: SplTokenStateCoder;
   readonly events: SplTokenEventsCoder;
+  readonly types: SplTokenTypesCoder;
 
   constructor(idl: Idl) {
     this.instruction = new SplTokenInstructionCoder(idl);
     this.accounts = new SplTokenAccountsCoder(idl);
     this.events = new SplTokenEventsCoder(idl);
     this.state = new SplTokenStateCoder(idl);
+    this.types = new SplTokenTypesCoder(idl);
   }
 }